Fix server rendering of Codeblock
