+++ packages/article-converter/package.json
@@ -26,9 +26,12 @@
     "es",
     "src"
   ],
+  "devDependencies": {
+    "@ndla/types-embed": "^0.1.3"
+  },
   "dependencies": {
-    "@ndla/code": "^2.1.13",
-    "@ndla/ui": "^34.1.2",
+    "@ndla/code": "^2.1.15",
+    "@ndla/ui": "^34.3.3",
     "html-react-parser": "^3.0.8",
     "lodash": "^4.17.20"
   },
