+++ packages/ndla-code/src/languageOptionts.ts
@@ -1,0 +1,115 @@
+export const languageOptions: Array<ICodeLangugeOption> = [
+  {
+    title: 'Bash',
+    format: 'bash',
+  },
+  {
+    title: 'C',
+    format: 'c',
+  },
+  {
+    title: 'CSharp',
+    format: 'csharp',
+  },
+  {
+    title: 'CSS',
+    format: 'css',
+  },
+  {
+    title: 'Diff',
+    format: 'diff',
+  },
+  {
+    title: 'HTML',
+    format: 'markup',
+  },
+  {
+    title: 'Ini',
+    format: 'ini',
+  },
+  {
+    title: 'Java',
+    format: 'java',
+  },
+  {
+    title: 'Javascript',
+    format: 'js',
+  },
+  {
+    title: 'JSON',
+    format: 'json',
+  },
+  {
+    title: 'JSX',
+    format: 'jsx',
+  },
+  {
+    title: 'Kotlin',
+    format: 'kotlin',
+  },
+  {
+    title: 'LUA',
+    format: 'lua',
+  },
+  {
+    title: 'Markdown',
+    format: 'markdown',
+  },
+  {
+    title: 'Matlab',
+    format: 'matlab',
+  },
+  {
+    title: 'NSIS',
+    format: 'nsis',
+  },
+  {
+    title: 'PHP',
+    format: 'php',
+  },
+  {
+    title: 'Powershell',
+    format: 'powershell',
+  },
+  {
+    title: 'Python',
+    format: 'python',
+  },
+  {
+    title: 'Ruby',
+    format: 'ruby',
+  },
+  {
+    title: 'Rust',
+    format: 'rust',
+  },
+  {
+    title: 'SQL',
+    format: 'sql',
+  },
+  {
+    title: 'Text',
+    format: 'text',
+  },
+  {
+    title: 'VHDL',
+    format: 'vhdl',
+  },
+  {
+    title: 'XML',
+    format: 'xml',
+  },
+];
+
+export interface ICodeLangugeOption {
+  title: string;
+  format: string;
+}
+
+export const getTitleFromFormat = (format: string) => {
+  const selectedLanguage = languageOptions.find((item: ICodeLangugeOption) => item.format === format);
+  if (selectedLanguage) {
+    return selectedLanguage.title;
+  }
+  return;
+};
